requirements.txt
pre-commit
python-box
pyyaml
requests
